♻️ refactor(cli): deduplicate anchor-walk, fix HOME startsWith check, guard filesystem root
Extract findSmithersAnchorDir into packages/smithers/src/findSmithersAnchorDir.js and
import it from both create.js and find-db.js, eliminating the copy-paste. Replace the
dir.length < home.length HOME comparison with dir.startsWith(home + "/") so symlinks
and unusual mounts don't produce false positives. Move the dir === root break to the top
of the findSmithersDb walk so /smithers.db is never added to allCandidates.
